refactor(printf): nano-safe formatting at every live printf site
Companion to the newlib-nano IROM lever (Package A, lever 1). Behaviour-
preserving on every env; no flag gates it, because the rewritten sites
produce the same strings under the full formatter they produce under nano.

What nano actually is on this toolchain, read out of the artifacts rather
than assumed (libc_nano.a's _svfprintf_r, which is what every snprintf and
LOG_ERR in this tree reaches -- the ROM only carries the stream variants):

  * length modifiers are exactly "hlL". %z, %j, %t, %hh and %ll are NOT
    parsed: the letter is handed to _printf_i as the conversion, printed
    literally, and the argument is left UNCONSUMED -- so every following
    argument shifts. A trailing %s then formats an integer as a pointer.
  * float conversions are "efgEFG" and dispatch through a WEAK undefined
    _printf_float. Nothing pulls nano-vfprintf_float.o in, so the symbol
    is null, the branch skips the conversion and advances the va_list by
    a double: "%.0f%%" renders as a bare "%".
  * argument-supplied width/precision ("%.*s") IS supported, contrary to
    the ledger note. The three sites below are rewritten anyway: the board
    name they print is not terminated at its own boundary, so a bounded
    copy is the honest spelling and it no longer depends on which
    formatter the env links.

Live sites fixed (live = LOG_ERR or a functional format; LOG_LEVEL=0
compiles LOG_DBG/LOG_INF out on the combo envs, and those ~25 float sites
are deliberately left alone -- the ini's nano entry carries the warning
that raising LOG_LEVEL revives them):

  * BaseTheme.cpp status-bar progress percent, drawn on every page turn in
    every theme. "%.0f%%" -> roundHalfToEven() + "%d%%". printf rounds to
    nearest and breaks a tie toward the even value; the helper reproduces
    that exactly, so the drawn string is unchanged for every value the
    four callers can pass (all non-negative, 0..100).
  * FontDownloadActivity::formatSize. "%.1f MB"/"%.0f KB" -> integer math
    on the scaled quotient. Both divisors are powers of two, so quotient
    and remainder are exact and the same ties-to-even rule reproduces the
    old output byte for byte (e.g. 1310720 B still prints "1.2 MB", not
    "1.3 MB"). The "%zu B" arm goes to %lu with it.
  * OtaUpdater's OTA asset suffix and the two board-mismatch LOG_ERRs ->
    board_tag::copyBoardName(), a bounded NUL-terminated copy, + "%s".
    Same truncation bound as the snprintf it replaces.
  * ImageToFramebufferDecoder's three %lld LOG_ERRs -> clampToLong() +
    %ld. Saturating, so a malformed header's int64 dimension still logs.
  * 27 %zu LOG_ERR sites across src/ and lib/ -> %lu with an explicit
    unsigned long cast (correct on the 32-bit target and on the 64-bit
    host harnesses alike). Epub.cpp's "CSS file too large (%zu > %zu): %s"
    was the worst of them: under nano the two unconsumed size_t arguments
    would have made %s dereference MAX_CSS_FILE_SIZE.

lib/expat's %8.2f / %*s / %0*lx traces are left as they are: all four sit
behind a runtime debug level that is zero on this device and an env var
that does not exist, exactly as the audit recorded.

Host suite: 351/354. The 3 not-run targets (InflateStreamTest,
SdCardFontTest, ChapterHtmlSlimParserTest) fail to configure their include
paths at this tip with or without this change -- verified by rebuilding
them against a stashed tree.

+++ lib/miniz/src/InflateStream.cpp
@@ -17,8 +17,10 @@ constexpr size_t STATE_ALIGNED = (sizeof(tinfl_decompressor) + 7) & ~size_t{7};
 void logAllocationFailure(const char* allocation, size_t bytes) {
   const auto heap = HalMemory::getDefaultHeap();
   const auto psram = HalMemory::getPsramHeap();
-  LOG_ERR("ZIP", "Inflate %s OOM (%zu bytes): heap %zu free/%zu max, PSRAM %zu free/%zu total", allocation, bytes,
-          heap.freeBytes, heap.largestBlockBytes, psram.freeBytes, psram.totalBytes);
+  LOG_ERR("ZIP", "Inflate %s OOM (%lu bytes): heap %lu free/%lu max, PSRAM %lu free/%lu total", allocation,
+          static_cast<unsigned long>(bytes), static_cast<unsigned long>(heap.freeBytes),
+          static_cast<unsigned long>(heap.largestBlockBytes), static_cast<unsigned long>(psram.freeBytes),
+          static_cast<unsigned long>(psram.totalBytes));
 }
 }  // namespace
